// apps/server/src/utils/enum.ts
export enum ResponseCode {
  SUCCESS = 200,
  SUCCESS_NEW_RESOURCE = 201,
  SUCCESS_WITHOUT_RESPONSE = 204,
  BAD_REQUEST = 400,
  UNAUTHORIZED = 401,
  NOT_FOUND = 404,
  FORBIDDEN = 403,
  UPGRADE_REQUIRED = 426,
  INTERNAL_SERVER = 500,
  MAINTENANCE = 503,
  TOKEN_INVALID = 498,
}
export enum MetaCode {
  SUCCESS = 1,
  FAIL = 0,
}